requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
pandas==2.2.2
python-multipart==0.0.9
pydantic==2.8.2
requests==2.32.3
python-dotenv==1.0.1
openai>=1.30.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.147.0